# lib/api-zod/src/generated/types/leaderboardResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LeaderboardEntry } from './leaderboardEntry';
import type { LeaderboardResponseScope } from './leaderboardResponseScope';

export interface LeaderboardResponse {
  scope: LeaderboardResponseScope;
  selfRank: number;
  entries: LeaderboardEntry[];
}