# lib/api-zod/src/generated/types/reviewDueResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReviewItemView } from './reviewItemView';

export interface ReviewDueResponse {
  dueCount: number;
  totalCount: number;
  items: ReviewItemView[];
}