# lib/api-zod/src/generated/types/shopResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ShopItem } from './shopItem';

export interface ShopResponse {
  coins: number;
  /** @nullable */
  equippedAvatar: string | null;
  /** @nullable */
  equippedTheme: string | null;
  items: ShopItem[];
}